Sage-main/sage-pro/requirements.txt
vllm==0.6.3
langgraph==0.2.34
langchain-core==0.3.15
fastapi==0.115.0
uvicorn[standard]==0.32.0
faiss-cpu==1.9.0
gudhi==3.10.1
sentence-transformers==3.2.0
structlog==24.4.0
pydantic==2.9.2
httpx==0.27.2
hypothesis==6.115.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
ruff==0.7.0
mypy==1.12.0
scipy==1.13.1
python-Levenshtein==0.26.0
bandit==1.7.10
semgrep==1.90.0
radon==6.0.1
pyperf==2.7.0
rich>=13.0.0
matplotlib>=3.7.0
numpy>=1.24.0
pyyaml>=6.0.0
datasets>=2.14.0
transformers>=4.34.0
sse-starlette>=1.6.0
autoawq>=0.1.0
# hyperfine (system) - assume available in environmentchromadb>=0.4.0
